requirements.txt
mcp>=1.0.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# Development dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0